Task-1/requirements.txt
numpy
pandas
python-pptx
Pillow
PyMuPDF
pytesseract
opencv-python
matplotlib
sentence-transformers
faiss-cpu
transformers